Fix: Fixed GUI files and kivy files inside 'src'.
- Inside the two GUI files, fixed an error with reading the "nuclear_species.txt" file. Implemented the python 'os' package to make sure the 'parent directory' is set correctly.

Note:
- The two files "backend_kivy.py" and "backend_kivyagg.py" are both technically unnecessary anymore but are just backups in case the user runs into import errors while trying to import "kivy.garden.matplotlib.

If the user is using the conda package management, they will most likely run into this import error since kivy-garden's way of installing matplotlib isn't compatible with conda.

The user can run "garden install matplotlib" in their terminal, which will install a working "kivy.garden.matplotlib' package in their user directory (NOT the conda environment directory!), but again, this is inconsistent with conda's package management style so the user might experience warnings from their IDE.

A workaround hack I found was to copy and paste the folder in the user directory into the conda environment's directory, but this is obviously not an ideal or permanent solution.

Conclusion: We will need to find a permanent solution with the kivy.garden.matplotlib import issue, then we can get rid of the two kivy files in our 'src' directory.