Major: fixed bug related to NuclearSpin vs ManySpins. Updated `simple_squeezing`.
This bug appeared while running the `simple_squeezing` demo notebook.
In the function `nuclear_system_setup` in `simulation.py`, the `spin_system` variable was being passed around always as an instance of the ManySpins class, although it was expected to be a NuclearSpin class.

Now the code has been updated so the `spin_system` is correctly an instance of NuclearSpin for a single spin system, and a ManySpins instance for multiple spins.

The `simple_squeezing.ipynb` notebook has been updated to fully run with no bugs.

(there was a refactoring of `ManySpins.spin` to `ManySpins.spins` happening at the same time, so remnants of this refactoring might appear in this commit)